requirements.txt
# Core LLM & Agents
crewai
langchain
langchain-community
langchain-google-genai
google-generativeai

# Web UI & Dev Tools
streamlit
python-dotenv

# Email & Formatting
markdown
pydantic

# Optional: Serper API support
requests